Use `@tanstack/react-query` for deposit/withdraw flow
Since we use the `@tanstack/react-query` library we no longer need the `useExecuteFunction` hook. Let's stay consistent use `useMutation` instead of `useExecuteFunction` in the rest of the code.
