+++ dapp/src/components/TransactionModal/ActiveStakingStep/DepositBTCModal.tsx
@@ -3,22 +3,25 @@ import {
   useActionFlowPause,
   useActionFlowTokenAmount,
   useAppDispatch,
-  useDepositBTCTransaction,
-  useExecuteFunction,
+  useConnector,
   useInvalidateQueries,
   useStakeFlowContext,
   useVerifyDepositAddress,
+  useWallet,
 } from "#/hooks"
 import { eip1193, logPromiseFailure } from "#/utils"
 import { PROCESS_STATUSES } from "#/types"
 import { setStatus, setTxHash } from "#/store/action-flow"
 import { ONE_SEC_IN_MILLISECONDS, queryKeysFactory } from "#/constants"
 import { useTimeout } from "@chakra-ui/react"
+import { useMutation } from "@tanstack/react-query"
 import WalletInteractionModal from "../WalletInteractionModal"
 
 const { userKeys } = queryKeysFactory
 
 export default function DepositBTCModal() {
+  const connector = useConnector()
+  const { address } = useWallet()
   const tokenAmount = useActionFlowTokenAmount()
   const { btcAddress, depositReceipt, stake } = useStakeFlowContext()
   const verifyDepositAddress = useVerifyDepositAddress()
@@ -38,12 +41,17 @@ export default function DepositBTCModal() {
     [dispatch],
   )
 
-  const handleStake = useExecuteFunction(stake, onStakeBTCSuccess, onError)
+  const { mutate: handleStake } = useMutation({
+    mutationKey: ["stake"],
+    mutationFn: stake,
+    onSuccess: onStakeBTCSuccess,
+    onError,
+  })
 
   const onDepositBTCSuccess = useCallback(
     (transactionHash: string) => {
       dispatch(setTxHash(transactionHash))
-      logPromiseFailure(handleStake())
+      handleStake()
     },
     [dispatch, handleStake],
   )
@@ -54,15 +62,41 @@ export default function DepositBTCModal() {
         handlePause()
       } else {
         onError()
+        console.error(error)
       }
     },
     [onError, handlePause],
   )
 
-  const { sendBitcoinTransaction, status } = useDepositBTCTransaction(
-    onDepositBTCSuccess,
-    onDepositBTCError,
-  )
+  const { mutate: sendBitcoinTransaction, status } = useMutation({
+    mutationKey: ["send-bitcoin"],
+    mutationFn: async ({
+      recipient,
+      amount,
+    }: {
+      recipient: string
+      amount: bigint
+    }) => {
+      if (!address) {
+        throw new Error("Bitcoin account was not connected.")
+      }
+
+      if (!connector) {
+        throw new Error("Connector was not defined.")
+      }
+      // @ts-expect-error adjust types to handle bitcoin wallet wrappers
+      const client = await connector.getClient()
+
+      const satoshis = Number(amount)
+
+      // @ts-expect-error adjust types to handle bitcoin wallet wrappers
+      // eslint-disable-next-line @typescript-eslint/no-unsafe-call, @typescript-eslint/no-unsafe-assignment
+      const txhash: string = await client.sendBitcoin(recipient, satoshis)
+      return txhash
+    },
+    onSuccess: onDepositBTCSuccess,
+    onError: onDepositBTCError,
+  })
 
   const handledDepositBTC = useCallback(async () => {
     if (!tokenAmount?.amount || !btcAddress || !depositReceipt) return
