+++ dapp/src/components/TransactionModal/ActiveStakingStep/StakingErrorModal/index.tsx
@@ -2,7 +2,6 @@ import React, { useCallback, useState } from "react"
 import {
   useActionFlowTxHash,
   useAppDispatch,
-  useExecuteFunction,
   useFetchActivities,
   useStakeFlowContext,
 } from "#/hooks"
@@ -10,6 +9,7 @@ import { PROCESS_STATUSES } from "#/types"
 import { logPromiseFailure } from "#/utils"
 import { setStatus } from "#/store/action-flow"
 import { UnexpectedErrorModalBase } from "#/components/UnexpectedErrorModal"
+import { useMutation } from "@tanstack/react-query"
 import ServerErrorModal from "./ServerErrorModal"
 import RetryModal from "./RetryModal"
 import LoadingModal from "../../LoadingModal"
@@ -24,7 +24,6 @@ export default function StakingErrorModal({
   const fetchActivities = useFetchActivities()
   const txHash = useActionFlowTxHash()
 
-  const [isLoading, setIsLoading] = useState(false)
   const [isServerError, setIsServerError] = useState(false)
 
   const onStakeBTCSuccess = useCallback(() => {
@@ -34,29 +33,21 @@ export default function StakingErrorModal({
 
   const onStakeBTCError = useCallback(() => setIsServerError(true), [])
 
-  const handleStake = useExecuteFunction(
-    stake,
-    onStakeBTCSuccess,
-    onStakeBTCError,
-  )
+  const { mutate: handleStake, status } = useMutation({
+    mutationKey: ["stake"],
+    mutationFn: stake,
+    onSuccess: onStakeBTCSuccess,
+    onError: onStakeBTCError,
+  })
 
-  const handleRetry = useCallback(async () => {
-    setIsLoading(true)
-    await handleStake()
-    setIsLoading(false)
-  }, [handleStake])
-
-  const handleRetryWrapper = useCallback(
-    () => logPromiseFailure(handleRetry()),
-    [handleRetry],
-  )
+  const isLoading = status === "pending"
 
   if (isServerError)
-    return <ServerErrorModal retry={handleRetryWrapper} isLoading={isLoading} />
+    return <ServerErrorModal retry={handleStake} isLoading={isLoading} />
 
   if (isLoading) return <LoadingModal />
 
-  if (txHash) return <RetryModal retry={handleRetryWrapper} />
+  if (txHash) return <RetryModal retry={handleStake} />
 
   return <UnexpectedErrorModalBase closeModal={closeModal} withCloseButton />
 }
