+++ dapp/src/hooks/index.ts
@@ -5,8 +5,6 @@ export * from "./useDetectThemeMode"
 export * from "./useSidebar"
 export * from "./useDocsDrawer"
 export * from "./useTransactionDetails"
-export * from "./useDepositBTCTransaction"
-export * from "./useExecuteFunction"
 export * from "./useStakeFlowContext"
 export * from "./useInitApp"
 export * from "./useCurrencyConversion"
